+++ packages/lint/src/index.ts
@@ -7,7 +7,7 @@ import preferPolarisComponent from './rules/prefer-polaris-component';
 
 const meta = {
   name: '@polaris/lint',
-  version: '0.4.1',
+  version: '0.5.0',
 } as const;
 
 const rules = {